AWS: grant stream service access to inbox (#21)
AWS: grant stream service access to inbox

The stream service should have read access to the inbox in order to
serve quarantined files back to a user, so that they can re-attempt
import. This is an upcoming feature of the API and stream service.

The examples for Azure and GCP do not provide granular IAM today, and
should not need changing.