Fix AWSexamples (#31)
These are the minimally required changes and pin the underlying modules
to get the examples to work as originally intended.

### Docs 

None 

### Description

Our AWS examples were failing due to breaking changes in the underlying
modules that we did not have pinned. This PR pins the modules to a known
working versions.

I tested this out by running against my AWS sandbox account and was able
to get terraform init, plan, and apply to work as expected.