Add outputs for Azure primary site (#33)
### Changelog
<!-- Write a one-sentence summary of the user-impacting change (API,
UI/UX, performance, etc) that could appear in a changelog. Write "None"
if there is no user-facing change -->

None

### Docs

<!-- Link to a Docs PR, tracking ticket in Linear, OR write "None" if no
documentation changes are needed. -->

None

### Description

Add outputs for the Azure primary site to make it easier to set cloud
credentials.

We've had these values of outputs of the respective modules, but they
don't get printed out after doing a `terraform apply`, so it makes it
hard to find the values. This change adds them to the root `outputs.tf`
file, so they can be viewed after an apply an accessed with `terraform
output <name>`.