Update service account to be query service (#35)
### Changelog
<!-- Write a one-sentence summary of the user-impacting change (API,
UI/UX, performance, etc) that could appear in a changelog. Write "None"
if there is no user-facing change -->

- aws: rename the stream-service service account to query-service

### Docs

None

<!-- Link to a Docs PR, tracking ticket in Linear, OR write "None" if no
documentation changes are needed. -->

### Description

This change PR reflects the changes made to the helm chart by renaming
`stream-service` to `query-service` for AWS sites.

This means the service account annotation in `values.yaml` will need to
be updated.